grserver/requirements.txt
fastapi
pytest
guardrails-ai==0.6.2
uvicorn
nltk
arize-otel
arize-phoenix
arize-phoenix-otel
openinference-instrumentation
openinference-instrumentation-guardrails
openinference-semantic-conventions